feat(fp8): convert multi_expert_pipeline weights to fp8 with bf16 cast at compute
Weight inputs changed from bf16 to float8_e4m3fn. In compute_tile,
each weight (w1/w3/w2) is cast to bf16 before jnp.dot — halving
VMEM weight buffer footprint while preserving bf16 MXU precision.

+++ tests/test_multi_expert_pipeline.py
@@ -92,7 +92,7 @@ class TestConfig:
 
     def test_dtype_is_bfloat16(self):
         assert self.config["dtype"] == "bfloat16"
-        assert self.config["weight_dtype"] == "bfloat16"
+        assert self.config["weight_dtype"] == "float8_e4m3fn"
 
     def test_act_fn_is_silu(self):
         assert self.config["act_fn"] == "silu"
